MM-43132: Run actions (#1160)
* Redux boilerplate

- Rename Action to ChannelAction in reducers, selectors and actions.
- Add run-related reducers, selectors, and actions, but don't use it yet.

* Refactor text are out of PatternedTextArea

The old PatternedTextArea, buried in the
src/components/backstage/playbook_edit/automation directory, contained
logic for the automation settings, used in the playbook editor, that
were not needed in the text area itself.
This commit promotes the text area element as the PatternedTextArea and
move this file to the base src/components directory. The old component
is now called WebhookSetting, which is just a thin wrapper around the
new PatternedTextAresa

* Refactor ActionsModal

- Refactor the channel-actions-specific logic out of ActionsModal, so
that the ActionsModal component is just a wrapper of GenericModal
preparing the field for a new RunActionsModal.
- Add a new ChannelActionsModal component that uses the now generic
ActionsModal. This new component is what was called ActionsModal, and
it's the one that is shown when clicking on ChannelActions in the
channel header menu entry.
- Simplify the Action and Trigger components: they now receive only the
props the need, not the whole ChannelAction object. Also, they no longer
are the owners of setting the titles strings, this is now done by the
caller that renders the component. All in all, they're simply UI
wrappers with little logic, which is now managed by the callers.

* Refactor broadcast channel selector

Decouple the inner broadcast channel selector from the automation
setting, so it can be reused out of the automation setting.

* Add the run actions modal

* Add button in run overview to open run actions

This commit also needed to refactor a bit the CopyIcon component, now
used not only for the copy button, but also for the run actions button.

* Add E2E tests

* make i18n-extract

* Fix E2E tests

* Allow saving empty actions

The server should not fail when saving an action with no keywords or an
empty category name, but simply treat those cases as no-op.

* Test saving empty actions, which is now allowed

There can be only one action of each type, so we need to create a whole
new test environment for each unit test. That's why this commit copies
the Setup and CreateBasic functions to each individual test. Otherwise,
some of the tests fail because a previous one already created an action
of the same type.

* Test that keywords can be removed

* Reset actions to empty state on channel switch

When the user switches between channels, the channel actions are fetched
to populate the actions modal. If there are actions, their state is set
to the fetched data. However, if there is some missing action, nothing
happened, as we only iterated through the returned actions. This commit
changes that, to ensure the state of all actions is reset: for those
fetched and for those that are not fetched.

Add an E2E test to check this case.

* Change the tooltip to Run Actions

* Send websocket event when run actions are updated

* Update webapp/src/client.ts

Co-authored-by: José Peso <trilopin@users.noreply.github.com>

* make i18n-extract

* Check that updating actions work

Retrieve the original and the updated action and verify that the payload
changes after making the update request.

* Create a new channel per test instead of a server

* Make linter happy

* Remove dead code

* Reset actions to the empty state *before* fetching

* Fix the modal rendering a previous step bug

Whenever the actions modal was saved, the underlying component
GenericModal was still executing onHide, even when
autoCloseOnConfirmButton was set to false. This happened because
GenericModal set the onExited prop to the passed onHide: this caused
onHide to be called after the modal had existed.

This created a race condition in ChannelActionsModal, which manages two
pieces of state to store the current state of the actions entered by the
user (let's call that ClientState) and the state of the actions stored
in the server (let's call that ServerState). This is supposed to work
like this:
1. When the modal is saved, ServerState is overwritten with whatever
ClientState had, as this is the information we are sending to the server.
2. When the modal is canceled (with the Cancel button or because the
user clicks outside of it), the ClientState is overwritter with whatever
ServerState had, as the changes are being discarded.

As onHide was incorrectly being called when onSave was executed, 2.
happened alongside with 1. In a best case scenario, this was useless, as
right after 1., ClientState and ServerState were exactly the same. In a
worst case scenario, this caused a bug, as ClientState was set to
ServerState *before* ServerState had the correct information.

This race condition was rare, and only happened on an E2E test, although
there, it was consistently failing. It was nearly impossible to
reproduce this manually.

* React code hygiene

Wrap the functions returned by useActionState in useCallback so that
they are re-computed only when their dependencies change. Otherwise,
they get generated on every render, which is wasteful.

Also, just in case, explicitly add the dependencies on these functions
to useEffect.

* Use icon from mdi instead of compass-icons

* Fix icon sizes

Co-authored-by: José Peso <trilopin@users.noreply.github.com>
Co-authored-by: Mattermod <mattermod@users.noreply.github.com>