update run modal description with broadcast info (#1162)
Brand new description in update run status modal.

The new description informs the user about how many channels and DM will be broadcasted with the update. A tooltip will be shown if the user does hover on the counters. Tooltips show up to 5 names and "and N others" in case we have more than 5.

If there are no channels or followers (DM), a different message will be shown in the description.

Additionally, the navigation is now done in the same tab (we force modal close if the user navigates).