MM-44241: ensure application/json
Apply the effects of https://mattermost.atlassian.net/browse/MM-40751 to Playbooks in the January ESR. Note that this is already fixed for newer versions of Playbooks that updated our webapp Client4 dependencies.

I opted to manually apply the patch here vs. bringing in all of the updated webapp dependencies.

Fixes: https://mattermost.atlassian.net/browse/MM-40751