Expose CSS property `right` in Dropdown component
This exposes an optional prop to change the value of the `right` CSS
property in the Dropdown component to adjust the horizontal position.
This replaces the old showOnRight prop, which was also removed from
those components that were explicitly using its default value (which
resulted in the default CSS `right: auto`).