npm install @mattermost/compass-icons@0.1.24
This explicitly adds compass-icons to our dependencies, moving from this
flow:

1. Make new release of compass-icons
2. Update webapp with the new version of compass-icons
3. Update Playbooks with the new version of webapp

to this:

1. Make new release of compass-icons
2. Update Playbooks with the new version of compass-icons

This may seem not too benefitial, but the time spent in 2. (open PR in
webapp, wait for dev review, wait for QA review (or at least communicate
that the PR doesn't need QA)) made us avoid that flow altogether and
come with alternative workarounds such as using mdiIcons. This,
hopefully, will make the process much faster and enjoyable.