MM-44719: Run Details Page - RHS Info Overview (#1265)
* Expose CSS property `right` in Dropdown component

This exposes an optional prop to change the value of the `right` CSS
property in the Dropdown component to adjust the horizontal position.
This replaces the old showOnRight prop, which was also removed from
those components that were explicitly using its default value (which
resulted in the default CSS `right: auto`).

* Add success/failure types to the toast component

This allows callers to specify which type of toast to render: if it's of
type ToastType.Failure, it will render a warning icon and the background
will be red; if it's of type ToastType.Success, it will render as
before: neutral background and a check icon.

* Expose more customizability in Following component

Let callers define the maximum number of users shown and whether or not
the help text 'n following' is shown.

* Add RHS info structure and RHS info overview

* make i18n-extract

* Make owner selector editable only for participants

* Encapsulate the following logic in a custom hook

* Simplify toggleFollow logic

* Fix toast message and make i18n-extract

* Make Playbook and Participants rows clickable

* npm install @mattermost/compass-icons@0.1.24

This explicitly adds compass-icons to our dependencies, moving from this
flow:

1. Make new release of compass-icons
2. Update webapp with the new version of compass-icons
3. Update Playbooks with the new version of webapp

to this:

1. Make new release of compass-icons
2. Update Playbooks with the new version of compass-icons

This may seem not too benefitial, but the time spent in 2. (open PR in
webapp, wait for dev review, wait for QA review (or at least communicate
that the PR doesn't need QA)) made us avoid that flow altogether and
come with alternative workarounds such as using mdiIcons. This,
hopefully, will make the process much faster and enjoyable.

* Use SVG icons from @mattermost/compass-icons

* Fix assignee dropdown in checklist hover menu

* Fix merge