Set scroll-padding-top when rendering run page
When navigating using a url like /playbooks/runs/<run_id>#summary, we
want the page to scroll to the element with id summary. This works out
of the box, but as the header is `position: fixed` we need to offset the
scroll value by its height, so that the element is not hidden under it.
Its height is 56px, but 10 additional pixels are added to give it a bit
more breathing room.