Refactor logic from createPlaybookRun  (#1284)
* Tidy up code in createPlaybookRun

This commit does not change any logic not internal structure, it just
reorders some statements so that the whole code is logically ordered and
refactor the block where we copy all relevant information from the
playbook to the run into another function.

* Move code updating playbook runs into PlaybookRun

The functions setPlaybookRunChecklist and copyPlaybookConfiguration
don't need to live in the API package. In fact, they could be plain
functions, but I decided to move them into the PlaybookRun struct to
group all functions related to the management of the data itself there.

* Use the same receiver name for all functions

i was used when the product was started because runs were called
incidents. r makes much more sense now.

* Removed numbering from comments