send message to channel when joined run has private channel (#1446)
When some user joins the run through participate but the channel is private, user is added to the run but not to the channel. Same happens when the user who add another new user to the run doesn't have access to the private channel.

For those cases, a message will be sent to the channel as playbooks bot.

cleanup:
- `leave` and `requestGetInvolved` REST run endpoints removed, also removed from client.ts
- `requestGetInvolved` service removed, part of it done in `participateActions`
- `leave` and `requestGetInvolved` removed from client


Graphql & client
I was tempted to add the functions addParticipants/removeParticipants to the client package, but I ended up thinking that graphql/schema is already the API (and we should not restrict/hide to API clients). I've just added a shortcut to handle the mutations more easily in our server tests (and not write them each time).


Fixes: https://github.com/mattermost/mattermost-plugin-playbooks/issues/1434