MM-47454: Introduce a highly visible feedback button (#1511)
The feedback button is an `a` that links out to a permalink, eventually landing on a Google Form. The current server version is passed as a query string parameter, though I note it's a Google Form specific parameter, meaning if we migrate to some other platform, we'll need to support this particular name.

Fixes: https://mattermost.atlassian.net/browse/MM-47454