remove channel coupling in ProfileSelector/AssignTo (#1538)
Evolve ProfileSelector / AssignTo to become decoupled from the channel.

Literals are now: run participants / not participating when groups are shown.
However it wasn't just a literals thing, we took as run members only those members in the channel. The significant change is to take the participant_ids list to decide whether they are members (and force ensureProfiles for them).
Some variables have been renamed in the ProfileSelector to avoid channel/run coupling.

The run list has a profile selector that is not linked to a run (but many). It was requesting runOwners without team_id which made the server answer with far more users than it should. It's fixed now.

Some real-time update problems have been solved too.