susbscribe to store for ephemeral weclome messages
We previously waited for websocket events issued by the client when marking a channel as viewed, but both the websocket message itself changed (to `multiple_channels_viewed`) and the semantics of when we trigger it (only if there is a new post to mark as viewed), so this approach stopped working in v9.0.

Instead, just listen for changes to the store channel and trigger the same flow.