Manually specify path to mattermost-redux files in Jest
I can't figure out why Jest can't find these files given other tools can
use the mappings in the package.json when other tools can. I have a hard
time telling if Jest should support that feature or not, but I've spent
enough time fighting with that to decide that I should leave it as-is
for now.