Replace incorrect usage of ActionFuncAsync
ActionFuncAsync shouldn't be be used in these places because the functions
passed through these props had already been dispatched, so the code using
them would only see the results of the action.

This is similar to a change we made a couple years ago in the web app to
standardize the types used for Redux actions.