Change type definitions for actions to be types instead of interfaces
I have no idea why, but having these be interfaces prevented us from
dispatching them because they couldn't be used as an UnknownAction. For some
reason, you can't pass an arbitrary interface as an UnknownAction because it
has an index signature, but TS will accept an arbitrary object just fine.